Move geopmbench.py into its own directory
