+++ integration/apps/geopmbench/Makefile.mk
@@ -29,14 +29,6 @@
 #  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
-EXTRA_DIST += integration/apps/apps.py \
-              integration/apps/__init__.py \
-              integration/apps/README.md \
+EXTRA_DIST += integration/apps/geopmbench/__init__.py \
+              integration/apps/geopmbench/geopmbench.py \
               # end
-
-include integration/apps/amg/Makefile.mk
-include integration/apps/geopmbench/Makefile.mk
-include integration/apps/hpcg/Makefile.mk
-include integration/apps/minife/Makefile.mk
-include integration/apps/nekbone/Makefile.mk
-include integration/apps/nasft/Makefile.mk
