+++ integration/experiment/monitor/run_monitor_dgemm.py
@@ -36,7 +36,7 @@ Run DGEMM with the monitor agent.
 '''
 
 from experiment.monitor import monitor
-from apps import geopmbench
+from apps.geopmbench import geopmbench
 
 
 if __name__ == '__main__':
