+++ integration/experiment/monitor/run_monitor_dgemm_tiny.py
@@ -37,7 +37,7 @@ Run a small DGEMM with the monitor agent.
 
 
 from experiment.monitor import monitor
-from apps import geopmbench
+from apps.geopmbench import geopmbench
 
 
 if __name__ == '__main__':
